Moves vex/inertial to vex/brain/inertial_sensor.
